requirements_new.txt
absl-py==0.11.0
appnope==0.1.2
astetik==1.11.1
astor==0.8.1
astunparse==1.6.3
backcall==0.2.0
cached-property==1.5.2
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chances==0.1.9
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
enum34==1.1.10
flatbuffers==1.12
gast==0.3.3
geonamescache==1.2.0
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==3.4.0
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.0
Keras==2.0.8
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kerasplotlib==0.1.6
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.4
mediapipe==0.8.2
numpy==1.19.3
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
pandas==1.1.5
parso==0.8.1
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
plaidbench==0.7.0
progress==1.5
prompt-toolkit==3.0.16
protobuf==3.14.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.8.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
scikit-learn==0.24.1
scipy==1.2.0
seaborn==0.11.1
six==1.15.0
sklearn==0.0
statsmodels==0.12.2
tensorboard==1.14.0
tensorboard-plugin-wit==1.8.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.56.2
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.26.3
wcwidth==0.2.5
Werkzeug==1.0.1
wrangle==0.6.7
wrapt==1.12.1
zipp==3.4.0
